feat: reorganize imports and update paths for product views and types
