+++ src/products/SmileIDBiometricKYCView.tsx
