+++ src/products/SmileIDBvnConsentScreenView.tsx
