+++ src/products/SmileIDDocumentVerificationEnhancedView.tsx
