+++ src/products/SmileIDEnhancedKyCView.tsx
