+++ src/products/SmileIDSmartSelfieAuthenticationEnhancedView.tsx
