+++ src/products/SmileIDSmartSelfieAuthenticationView.tsx
