+++ src/products/SmileIDSmartSelfieEnrollmentEnhancedView.tsx
