+++ src/products/SmileIDSmartSelfieEnrollmentView.tsx
